backend/requirements.txt
fastapi
pydantic-settings
sqlalchemy
psycopg2-binary
mangum
python-dotenv
langchain-aws
langchain-community
langchain-postgres
pgvector
boto3
openai
